requirements.txt

torch>=2.2
torchvision>=0.17
torchaudio>=2.2
pyyaml>=6.0
tqdm>=4.66
numpy>=1.26
scikit-learn>=1.4
Pillow>=10.0
matplotlib>=3.8
